Controller class refactor
