+++ backend/automated-tests/src/main/kotlin/br/com/zup/beagle/automatedtests/controllers/GeneralController.kt
@@ -24,23 +24,12 @@ import br.com.zup.beagle.automatedtests.constants.SAFE_AREA_ENDPOINT
 import br.com.zup.beagle.automatedtests.constants.SCREEN_ACTION_CLICK_ENDPOINT
 import br.com.zup.beagle.automatedtests.constants.SCREEN_SAFE_AREA_FALSE_ENDPOINT
 import br.com.zup.beagle.automatedtests.constants.SCREEN_SAFE_AREA_TRUE_ENDPOINT
-import br.com.zup.beagle.automatedtests.constants.SEND_REQUEST_ACTION_ENDPOINT
-import br.com.zup.beagle.automatedtests.builders.SendRequestScreenBuilder
 import br.com.zup.beagle.automatedtests.builders.ClickActionScreenBuilder
 import br.com.zup.beagle.automatedtests.builders.AnalyticsScreenBuilder
 import br.com.zup.beagle.automatedtests.builders.SafeAreaScreenBuilder
 import br.com.zup.beagle.automatedtests.builders.BuilderKotlinDslScreenBuilder
 import br.com.zup.beagle.automatedtests.builders.ExpressionEscapingScreenBuilder
 import br.com.zup.beagle.automatedtests.builders.AccessibilityScreenBuilder
-import br.com.zup.beagle.automatedtests.builders.ActionNotRegisteredScreenBuilder
-import br.com.zup.beagle.automatedtests.builders.AlertScreenBuilder
-import br.com.zup.beagle.automatedtests.builders.SetContextScreenBuilder
-import br.com.zup.beagle.automatedtests.builders.ConfirmScreenBuilder
-import br.com.zup.beagle.automatedtests.constants.ALERT_SCREEN_ENDPOINT
-import br.com.zup.beagle.automatedtests.constants.NO_ACTION_SCREEN_ENDPOINT
-import br.com.zup.beagle.automatedtests.constants.SET_CONTEXT_SCREEN_ENDPOINT
-import br.com.zup.beagle.automatedtests.constants.CONFIRM_SCREEN_ENDPOINT
-import br.com.zup.beagle.automatedtests.constants.NO_ACTION_SCREEN_ENDPOINT
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
 
@@ -70,21 +59,6 @@ class GeneralController {
 
     @GetMapping(ACCESSIBILITY_ENDPOINT)
     fun getAccessibilityScreen() = AccessibilityScreenBuilder.build()
-
-    @GetMapping(SEND_REQUEST_ACTION_ENDPOINT)
-    fun getSendRequestScreen() = SendRequestScreenBuilder.build()
-
-    @GetMapping(ALERT_SCREEN_ENDPOINT)
-    fun getAlertScreen() = AlertScreenBuilder.build()
-
-    @GetMapping(NO_ACTION_SCREEN_ENDPOINT)
-    fun getActionNotRegisteredScreen() = ActionNotRegisteredScreenBuilder.build()
-
-    @GetMapping(SET_CONTEXT_SCREEN_ENDPOINT)
-    fun getSetContextScreen() = SetContextScreenBuilder.build()
-
-    @GetMapping(CONFIRM_SCREEN_ENDPOINT)
-    fun getConfirmScreen() = ConfirmScreenBuilder.build()
 }
